+++ hrk_nasa_apis_test/tool/ci.sh
@@ -10,6 +10,8 @@ dart format --output none --set-exit-if-changed .
 
 dart analyze --fatal-infos
 
+tool/test.sh
+
 dart pub global activate pana
 dart pub global run pana --no-warning